+++ test/index.js
@@ -3,10 +3,13 @@ describe('General functionality', () => {
   require('./general')
 })
 
+describe('CORS and PNA protection', () => {
+  require('./cors')
+})
+
 describe('Senso', () => {
   require('./senso')
 })
-
 
 describe('RFID', () => {
   require('./rfid')