+++ Changelog.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- Deny WebSocket connections from unknown origins
 - Let firmware update command look for bootloader if no Senso in regular mode is found
 - Update build system and development environment